Adds method to organisations Api to add users to the org
