+++ app/controllers/carto/api/organizations_controller.rb
@@ -5,9 +5,10 @@ require_relative './user_presenter'
 module Carto
   module Api
     class OrganizationsController < ::Api::ApplicationController
+      include OrganizationsHelper
       include PagedSearcher
 
-      ssl_required :users
+      ssl_required
 
       before_filter :load_organization, :load_group
 
@@ -26,18 +27,23 @@ module Carto
         }, total_user_entries: total_user_entries, total_entries: users.count })
       end
 
-      def load_organization
-        @organization = Carto::Organization.where(id: params[:id]).first
-        render_jsonp({}, 401) and return if @organization.nil?
-      end
+      def create_user
+        user = ::User.new(create_user_params)
+        user.organization_id = @organization.id
 
-      def load_group
-        if params[:group_id]
-          @group = @organization.groups.find(params[:group_id])
-          render_jsonp({ errors: "No #{params[:group_id]} at #{@organization.id}" }, 404) and return unless @group
+        if user.save
+          render_jsonp status_message(200, Carto::Api::UserPresenter.new(user).to_poro)
+        else
+          render_jsonp status_message(410, user.errors.full_messages)
         end
       end
 
+      private
+
+      # TODO: Use native strong params when in Rails 4+
+      def create_user_params
+        permit(:email, :username, :password, :password_confirmation, :quota_in_bytes)
+      end
     end
   end
 end
